+++ ktrain/tests/test_regression.py
@@ -33,7 +33,7 @@ class TestRegression(TestCase):
         hist  = bostonhousing()
         min_loss = min(hist.history['val_loss'])
         print(min_loss)
-        self.assertLess(min_loss, 40)
+        self.assertLess(min_loss, 45)
 
 if __name__ == "__main__":
     main()