+++ ktrain/text/preprocessor.py
@@ -1,22 +1,76 @@
+from ..imports import *
+from .. import utils as U
 from ..preprocessor import Preprocessor
-from keras.preprocessing import image
-from keras.preprocessing import sequence
+
+
+#BERT_PATH = os.path.join(os.path.dirname(os.path.abspath(localbert.__file__)), 'uncased_L-12_H-768_A-12')
+BERT_URL = 'https://storage.googleapis.com/bert_models/2018_10_18/uncased_L-12_H-768_A-12.zip'
+
+def get_bert_zip():
+    return os.path.split(BERT_URL)[-1]
+
+def get_bert_folder():
+    return os.path.splitext(get_bert_zip())[0]
+
+def get_ktrain_data():
+    home = os.path.expanduser('~')
+    ktrain_data = os.path.join(home, 'ktrain_data')
+    if not os.path.isdir(ktrain_data):
+        os.mkdir(ktrain_data)
+    return ktrain_data
+
+def get_bert_path():
+    ktrain_data = get_ktrain_data()
+    bert_path = os.path.join(ktrain_data, get_bert_folder())
+    zip_fname = get_bert_zip()
+    zip_fpath = os.path.join(ktrain_data, zip_fname)
+    if not os.path.isdir(bert_path) or \
+       not os.path.isfile(os.path.join(bert_path, 'bert_config.json')) or\
+       not os.path.isfile(os.path.join(bert_path, 'bert_model.ckpt.data-00000-of-00001')) or\
+       not os.path.isfile(os.path.join(bert_path, 'bert_model.ckpt.index')) or\
+       not os.path.isfile(os.path.join(bert_path, 'bert_model.ckpt.meta')) or\
+       not os.path.isfile(os.path.join(bert_path, 'vocab.txt')):
+        # download zip
+        print('downloading pretrained BERT model...')
+        #urllib.request.urlretrieve(BERT_URL, filename=zip_fpath)
+        #wget.download(BERT_URL, zip_fpath)
+        U.download(BERT_URL, zip_fpath)
+        #urllib.request.urlretrieve(BERT_URL, filename=zip_fpath)
+
+        # unzip
+        print('\nextracting pretrained BERT model...')
+        with zipfile.ZipFile(zip_fpath, 'r') as zip_ref:
+            zip_ref.extractall(ktrain_data)
+        print('done.\n')
+    return bert_path
+
+
+
+def bert_tokenize(docs, tokenizer, maxlen, verbose=1):
+    indices = []
+    mb = master_bar(range(1))
+    for i in mb:
+        for doc in progress_bar(docs, parent=mb):
+            ids, segments = tokenizer.encode(doc, max_len=maxlen)
+            indices.append(ids)
+        if verbose: mb.write('done.')
+    zeros = np.zeros_like(indices)
+    return [np.array(indices), np.array(zeros)]
+
+
 class TextPreprocessor(Preprocessor):
     """
-    Text preprocessing
+    Text preprocessing base class
     """
 
-    def __init__(self, tok, tok_dct, classes, maxlen, ngram_range=1):
+    def __init__(self, maxlen, classes):
 
-        self.tok = tok
-        self.tok_dct = tok_dct
         self.c = classes
         self.maxlen = maxlen
-        self.ngram_range = ngram_range
 
 
     def get_preprocessor(self):
-        return (self.tok, self.tok_dct)
+        raise NotImplementedError
 
 
     def get_classes(self):
@@ -24,10 +78,36 @@ class TextPreprocessor(Preprocessor):
 
 
     def preprocess(self, texts):
-        texts = self.tok.texts_to_sequences(texts)
-        texts = self.add_ngram(texts)
-        texts = sequence.pad_sequences(texts, maxlen=self.maxlen)
-        return texts
+        raise NotImplementedError
+
+
+    def undo(self, doc):
+        """
+        undoes preprocessing and returns raw data by:
+        converting a list or array of Word IDs back to words
+        """
+        raise NotImplementedError
+
+
+class StandardTextPreprocessor(TextPreprocessor):
+    """
+    Standard text preprocessing
+    """
+
+    def __init__(self, maxlen, max_features, classes=[], ngram_range=1):
+        super().__init__(maxlen, classes)
+        self.tok = None
+        self.tok_dct = {}
+        self.max_features = max_features
+        self.ngram_range = ngram_range
+
+
+    def get_preprocessor(self):
+        return (self.tok, self.tok_dct)
+
+
+    def preprocess(self, texts):
+        return self.preprocess_test(texts, verbose=0)[0]
 
 
     def undo(self, doc):
@@ -39,15 +119,98 @@ class TextPreprocessor(Preprocessor):
         return " ".join([dct[wid] for wid in doc if wid != 0 and wid in dct])
 
 
-    def ngram_count(self):
-        if not self.tok_dct: return 1
-        s = set()
-        for k in self.tok_dct.keys():
-            s.add(len(k))
-        return max(list(s))
+    def preprocess_train(self, train_text, y_train, verbose=1):
+        """
+        preprocess training set
+        """
+        # extract vocabulary
+        self.tok = Tokenizer(num_words=self.max_features)
+        self.tok.fit_on_texts(train_text)
+        U.vprint('Word Counts: {}'.format(len(self.tok.word_counts)), verbose=verbose)
+        U.vprint('Nrows: {}'.format(len(train_text)), verbose=verbose)
+
+        # convert to word IDs
+        x_train = self.tok.texts_to_sequences(train_text)
+        U.vprint('{} train sequences'.format(len(x_train)), verbose=verbose)
+        U.vprint('Average train sequence length: {}'.format(np.mean(list(map(len, x_train)), dtype=int)), verbose=verbose)
+
+        # add ngrams
+        x_train = self._fit_ngrams(x_train, verbose=verbose)
+
+        # pad sequences
+        x_train = sequence.pad_sequences(x_train, maxlen=self.maxlen)
+
+        # transform y
+        if len(y_train.shape) == 1:
+            y_train = to_categorical(y_train)
+        U.vprint('x_train shape: ({},{})'.format(x_train.shape[0], x_train.shape[1]), verbose=verbose)
+        U.vprint('y_train shape: ({},{})'.format(y_train.shape[0], y_train.shape[1]), verbose=verbose)
+
+        # return
+        return (x_train, y_train)
 
 
-    def add_ngram(self, sequences):
+    def preprocess_test(self, test_text, y_test=None, verbose=1):
+        """
+        preprocess validation or test dataset
+        """
+        if self.tok is None:
+            raise Exception('No tokenizer fitted. Did you run preprocess_train first?')
+
+        # convert to word IDs
+        x_test = self.tok.texts_to_sequences(test_text)
+        U.vprint('{} test sequences'.format(len(x_test)), verbose=verbose)
+        U.vprint('Average test sequence length: {}'.format(np.mean(list(map(len, x_test)), 
+                                                                 dtype=int)), verbose=verbose)
+
+        # add n-grams
+        x_test = self._add_ngrams(x_test, mode='test', verbose=verbose)
+
+
+        # pad sequences
+        x_test = sequence.pad_sequences(x_test, maxlen=self.maxlen)
+        U.vprint('x_test shape: ({},{})'.format(x_test.shape[0], x_test.shape[1]), verbose=verbose)
+
+        # convert y
+        if y_test is not None:
+            if len(y_test.shape) == 1:
+                y_test = to_categorical(y_test)
+            U.vprint('y_test shape: ({},{})'.format(y_test.shape[0], y_test.shape[1]), verbose=verbose)
+
+        # return
+        return (x_test, y_test)
+
+
+
+    def _fit_ngrams(self, x_train, verbose=1):
+        self.tok_dct = {}
+        if self.ngram_range < 2: return x_train
+        U.vprint('Adding {}-gram features'.format(self.ngram_range), verbose=verbose)
+        # Create set of unique n-gram from the training set.
+        ngram_set = set()
+        for input_list in x_train:
+            for i in range(2, self.ngram_range + 1):
+                set_of_ngram = self._create_ngram_set(input_list, ngram_value=i)
+                ngram_set.update(set_of_ngram)
+
+        # Dictionary mapping n-gram token to a unique integer.
+        # Integer values are greater than max_features in order
+        # to avoid collision with existing features.
+        start_index = self.max_features + 1
+        token_indice = {v: k + start_index for k, v in enumerate(ngram_set)}
+        indice_token = {token_indice[k]: k for k in token_indice}
+        self.tok_dct = token_indice
+
+        # max_features is the highest integer that could be found in the dataset.
+        self.max_features = np.max(list(indice_token.keys())) + 1
+        U.vprint('max_features changed to %s with addition of ngrams' % (self.max_features), verbose=verbose)
+
+        # Augmenting x_train with n-grams features
+        x_train = self._add_ngrams(x_train, verbose=verbose, mode='train')
+        return x_train
+
+
+    def _add_ngrams(self, sequences, verbose=1, mode='test'):
         """
         Augment the input list of list (sequences) by appending n-grams values.
         Example: adding bi-gram
@@ -63,8 +226,92 @@ class TextPreprocessor(Preprocessor):
                     if ngram in token_indice:
                         new_list.append(token_indice[ngram])
             new_sequences.append(new_list)
+        U.vprint('Average {} sequence length with ngrams: {}'.format(mode,
+            np.mean(list(map(len, new_sequences)), dtype=int)), verbose=verbose)    
         return new_sequences
 
 
 
+    def _create_ngram_set(self, input_list, ngram_value=2):
+        """
+        Extract a set of n-grams from a list of integers.
+        >>> create_ngram_set([1, 4, 9, 4, 1, 4], ngram_value=2)
+        {(4, 9), (4, 1), (1, 4), (9, 4)}
+        >>> create_ngram_set([1, 4, 9, 4, 1, 4], ngram_value=3)
+        [(1, 4, 9), (4, 9, 4), (9, 4, 1), (4, 1, 4)]
+        """
+        return set(zip(*[input_list[i:] for i in range(ngram_value)]))
+
+
+        def ngram_count(self):
+            if not self.tok_dct: return 1
+            s = set()
+            for k in self.tok_dct.keys():
+                s.add(len(k))
+            return max(list(s))
+
+
+class BERTPreprocessor(TextPreprocessor):
+    """
+    text preprocessing for BERT model
+    """
+
+    def __init__(self, maxlen, max_features, classes=[], ngram_range=1):
+
+        if maxlen > 512: raise ValueError('BERT only supports maxlen <= 512')
+
+        super().__init__(maxlen, classes)
+        vocab_path = os.path.join(get_bert_path(), 'vocab.txt')
+        token_dict = {}
+        with codecs.open(vocab_path, 'r', 'utf8') as reader:
+            for line in reader:
+                token = line.strip()
+                token_dict[token] = len(token_dict)
+        tokenizer = BERT_Tokenizer(token_dict)
+        self.tok = tokenizer
+        self.tok_dct = token_dict
+
+
+    def get_preprocessor(self):
+        return (self.tok, self.tok_dct)
+
+
+
+    def preprocess(self, texts):
+        return self.preprocess_test(texts, verbose=0)[0]
+
+
+    def undo(self, doc):
+        """
+        undoes preprocessing and returns raw data by:
+        converting a list or array of Word IDs back to words
+        """
+        dct = self.tok_dct
+        return " ".join([dct[wid] for wid in doc if wid != 0 and wid in dct])
+
+
+    def preprocess_train(self, texts, y=None, mode='train', verbose=1):
+        """
+        preprocess training set
+        """
+        U.vprint('preprocessing %s...' % (mode), verbose=verbose)
+        x = bert_tokenize(texts, self.tok, self.maxlen, verbose=verbose)
+        if y is not None:
+            if len(y.shape) == 1:
+                y = to_categorical(y)
+            #U.vprint('\ty shape: ({},{})'.format(y.shape[0], y.shape[1]), verbose=verbose)
+        return (x, y)
+
+
+
+    def preprocess_test(self, texts, y=None, mode='test', verbose=1):
+        return self.preprocess_train(texts, y=y, mode=mode, verbose=verbose)
+
+
+
+
+
+# preprocessors
+TEXT_PREPROCESSORS = {'standard': StandardTextPreprocessor,
+                      'bert': BERTPreprocessor}